fix: added missing WinRM password in qemu build
Qemu build for Windows was not working as WinRM password was missing.

In addition all VirtIO drivers are now moved to the installation CD
as packer overwrites all disk args if one is manually adding
`qemuargs` to the build file containing `-drive`. Hence, attaching
an additional CD for VirtIO is not possible when at the same time
`cd_contents` is used as this generates a unique name for the
cd attached.

Finally updated the Qemu Plugin to newest version as well as setting
the minimum required version of packer for all builds to the current
one used (as these are tested).